backend/requirements.txt
# Squadbox Backend Requirements
# Database: gdiba2_squadbox
# Server: gdiba2.ssh.tb-hosting.com

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pymongo==4.6.0

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# AI/LLM
openai==1.3.7
ollama==0.1.7

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
markdown==3.5.1

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
